+++ python/GafferImageTest/__init__.py
@@ -110,6 +110,7 @@ from .DeleteViewsTest import DeleteViewsTest
 from .CopyViewsTest import CopyViewsTest
 from .AnaglyphTest import AnaglyphTest
 from .LookTransformTest import LookTransformTest
+from .OpenColorIOAlgoTest import OpenColorIOAlgoTest
 
 if __name__ == "__main__":
 	import unittest